extract shared timeline types and utilities
